requirements.txt
rich>=13.7.0
ollama>=0.1.7
watchdog>=3.0.0
python-dateutil>=2.8.2
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

